Convert math structures back to float, see issue #2653. This gives us a small but noticeable performance boost of about 10-20% and smaller memory footprint, and it leaves the door open to using SIMD instructions more effectively sometime in the future.

+++ include/iorthoview.h
@@ -5,7 +5,7 @@
 
 template<typename Element>
 class BasicVector3;
-typedef BasicVector3<double> Vector3;
+typedef BasicVector3<float> Vector3;
 
 // Possible types of the orthogonal view window
 enum EViewType
