+++ include/itransformable.h
@@ -11,9 +11,9 @@ enum TransformModifierType
 
 // Forward-declare the math objects
 template<typename Element> class BasicVector3;
-typedef BasicVector3<double> Vector3;
+typedef BasicVector3<float> Vector3;
 template<typename Element> class BasicVector4;
-typedef BasicVector4<double> Vector4;
+typedef BasicVector4<float> Vector4;
 typedef Vector4 Quaternion;
 
 class ITransformable
